+++ src/ts/Ana/Ana.ts
@@ -4,6 +4,7 @@ import { RenderDictionary } from '../types'
 import { AttributeValuesDictionary, MatchFunctionsDictionary } from '../types'
 import * as utils from '../utils'
 import { rLayout } from '../Components/Atoms/Layout/Layout'
+import { rSurface } from '../Components/Atoms/Surface/Surface'
 // import { rDisplay } from '../Components/Atoms/Display/Display'
 // import { rTitle } from '../Components/Atoms/Title/Title'
 // import { rTestpage } from '../Components/Organisms/Testpage/Testpage'
@@ -94,7 +95,8 @@ export class Ana implements iAna {
     HTMLElement.prototype.setAttributes = this.setAttributes
     let elements = getElements(this.configuration)
     let atoms = {
-      Layout: rLayout(elements, this.configuration)
+      Layout: rLayout(elements, this.configuration),
+      Surface: rSurface(elements, this.configuration)
       //   Display: rDisplay(elements, this.configuration),
       //   Title: rTitle(elements, this.configuration),
     }