+++ src/ts/Components/Particles/Particles.ts
@@ -1,16 +1,15 @@
-export function textContrastClass(contrast?:string) {
-  return {
-    'txt-cont-hst': contrast === 'highest',
-    'txt-cont-h': contrast === 'high',
-    'txt-cont-m': contrast === 'mid',
-    'txt-cont-l': contrast === 'low',
-    'txt-cont-lst': contrast === 'lowest',
-  }
+export function ColorContrastClass(color?:string, contrast?:string, type?: 'txt' | 'bg') {
+return [`${color}-${type}`, {
+    'cont-hst': contrast === 'highest',
+    'cont-h': contrast === 'high',
+    'cont-m': contrast === 'mid',
+    'cont-l': contrast === 'low',
+    'cont-lst': contrast === 'lowest',
+  }]
 }
 
 export function LayoutClass(type?:string, margin?:string) {
   return [{
-
     'a-Layout-full': type === 'full',
     'a-Layout-center': type === 'center',
     'a-Layout-focus': type === 'focus',
@@ -19,7 +18,6 @@ export function LayoutClass(type?:string, margin?:string) {
     'a-Layout-quarters': type === 'quarters',
     'a-Layout-right': type === 'right',
     'a-Layout-left': type === 'left'
-
   }, {
     'a-Layout-margin-full': margin === 'full',
     'a-Layout-margin-single': margin === 'single',